requirements.txt
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=2.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langsmith>=0.2.0
tavily-python>=0.5.0
streamlit>=1.38.0
python-dotenv>=1.0.0
